Identify and move shared data file
