+++ src/data/bot_impact/assess_all_days.csv
